check if no visual scheme defined in previous commit
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@818782 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/content/src/org/ofbiz/content/data/DataResourceWorker.java
@@ -694,7 +694,7 @@ public class DataResourceWorker  implements org.ofbiz.widget.DataResourceWorkerI
                      String docbookStyleLocation = docbookStyleSheets.get(0);
                      sourceFileLocation = new File(System.getProperty("ofbiz.home")+"/themes"+docbookStyleLocation);
                  }
-                 if(sourceFileLocation.exists()){
+                 if(sourceFileLocation != null && sourceFileLocation.exists()){
                          UtilMisc.copyFile(sourceFileLocation,targetFileLocation);
                  }
                     // get the template data for rendering
